+++ project-base/src/Shopsys/ShopBundle/Model/Newsletter/NewsletterRepository.php
@@ -30,19 +30,22 @@ class NewsletterRepository
      * @param int $domainId
      * @return bool
      */
-    public function existsSubscribedEmail($email, $domainId)
+    public function existsSubscribedEmailByDomainId($email, $domainId)
     {
         return $this->getNewsletterSubscriberRepository()->findOneBy(['email' => $email, 'domainId' => $domainId]) !== null;
     }
 
     /**
+     * @param int $domainId
      * @return \Doctrine\ORM\Internal\Hydration\IterableResult
      */
-    public function getAllEmailsDataIterator()
+    public function getAllEmailsDataIteratorByDomain($domainId)
     {
         $query = $this->getNewsletterSubscriberRepository()
             ->createQueryBuilder('ns')
             ->select('ns.email, ns.createdAt')
+            ->where('ns.domainId = :domainId')
+            ->setParameter('domainId', $domainId)
             ->getQuery();
 
         return $query->iterate(null, AbstractQuery::HYDRATE_SCALAR);