+++ project-base/CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - personal data 
     - orders  
     - newsletter subscirbtion
+- Newsletter subscriber is distinguished by domain (@stanoMilan) 
+    - emails are exported by domain
+
+
 
 ### Changed
 - cache is cleared before PHPUnit tests only when run via [Phing targets](docs/introduction/phing-targets.md), not when run using `phpunit` directly (@PetrHeinz)